以降の変換処理で再利用するため、get_property_typeを外出し
